Add files via upload
Wikishop online store launches a new service. Now users can edit and supplement product descriptions, as in wiki communities. That is, clients offer their edits and comment on the changes of others. The store needs a tool that will search for toxic comments and send them for moderation.
In this project, we will train a machine learning model to classify comments into positive and negative. Our goal will be to reach F1 at least 0.75.